requirements.txt
# Django & Django Ninja
Django==5.0.1
django-ninja==1.1.0

# Database
psycopg[binary]>=3.2.2

# Authentication & Security
PyJWT==2.8.0
python-decouple==3.8
django-cors-headers==4.3.1

# File Upload & Images
Pillow==11.0.0

# Development
pytest==7.4.4
pytest-django==4.7.0
black==24.1.1
flake8==7.0.0

# Utils
python-dateutil==2.8.2
